Core FF file Prism.* split off
