+++ Core/HardParticle/IFormFactorPrism.h
@@ -15,26 +15,10 @@
 #ifndef BORNAGAIN_CORE_HARDPARTICLE_IFORMFACTORPRISM_H
 #define BORNAGAIN_CORE_HARDPARTICLE_IFORMFACTORPRISM_H
 
-#include "Core/HardParticle/PolyhedralComponents.h"
-#include "Core/HardParticle/PolyhedralTopology.h"
 #include "Core/Scattering/IFormFactorBorn.h"
 #include <memory>
 
-class BA_CORE_API_ Prism
-{
-public:
-    Prism() = delete;
-    Prism(const Prism&) = delete;
-    Prism(bool symmetry_Ci, double height, const std::vector<kvector_t>& vertices);
-    double area() const;
-    const std::vector<kvector_t>& vertices(); //! needed for topZ, bottomZ computation
-    complex_t evaluate_for_q(const cvector_t& q) const;
-    // complex_t evaluate_centered(const cvector_t& q) const;
-private:
-    std::unique_ptr<PolyhedralFace> m_base;
-    double m_height;
-    std::vector<kvector_t> m_vertices; //! for topZ, bottomZ computation only
-};
+class Prism;
 
 //! A prism with a polygonal base, for form factor computation.
 
@@ -43,6 +27,7 @@ class BA_CORE_API_ IFormFactorPrism : public IFormFactorBorn
 public:
     IFormFactorPrism() = default;
     IFormFactorPrism(const NodeMeta& meta, const std::vector<double>& PValues);
+    ~IFormFactorPrism();
 
     double bottomZ(const IRotation& rotation) const override final;
     double topZ(const IRotation& rotation) const override final;
